Convert money fields to decimals

+++ app/Models/Order.php
@@ -32,7 +32,7 @@ class Order extends Model
     {
         return [
             'delivery_date' => 'datetime',
-            'total_cost' => 'integer',
+            'total_cost' => 'float',
             'total_transit_days' => 'integer',
             'status' => OrderState::class,
             'delivery_day' => 'integer',
